### 2018-08-15 #895 @whedon
[PRE REVIEW]: idpflex: Analysis of Intrinsically Disordered Proteins by Comparing Simulations to Small Angle Scattering Experiments
**Submitting author:** @jmborr (<a href="http://orcid.org/0000-0002-0866-8158">Jose M. Borreguero</a>)
**Repository:** <a href="https://github.com/jmborr/idpflex.git" target ="_blank">https://github.com/jmborr/idpflex.git</a>
**Version:** v0.1.5
