### 2018-08-15 #895 @arfon
@jmborr - it looks like the `orcid` fields in your `paper.md` file aren't formed correctly. Instead of
```
ORCID 0000-0002-0866-8158
```
they should be

```
orcid: 0000-0002-0866-8158
```

In addition, your paper doesn't look to be complete?
