### 2018-08-15 01e39a9
Refs #70 lowercase paper
