### 2018-08-15 #895 @jmborr
@arfon First, let me thank you for taking the time to review the submission. Second, let me apologize for a silly mistake. I prepared the submission to JOSS in a dedicated branch and I became engrossed to the point that I forgot to merge it into master :blush:
I've just merged the branch, I believe the paper is now complete.
