### 2018-08-15 #895 @arfon
Hi @jmborr, please note the `orcid` fields need colons separating the key/value pairs.

i.e. change this `orcid 0000-0002-0866-8158` to this `orcid: 0000-0002-0866-8158`
