### 2018-08-16 aefbc0c
Update paper.md
