### 2018-08-16 79c7bf1
correct reference formatting
