+++ paper/paper.bib
@@ -2,7 +2,7 @@
     Author = {Curtis, Joseph E. and Raghunandan, Sindhu and Nanda, Hirsh and Krueger, Susan},
     Title = {{SASSIE: A program to study intrinsically disordered biological molecules and macromolecular ensembles using experimental scattering restraints}},
     Journal = {{Computer Physics Communications}},
-    Year = {{2012}},
+    Year = {2012},
     Volume = {{183}},
     Number = {{2}},
     Pages = {{382-389}},
@@ -20,7 +20,7 @@ Author = {Svergun, D and Barberato, C and Koch, MHJ},
 Title = {{CRYSOL - A program to evaluate x-ray solution scattering of biological
    macromolecules from atomic coordinates}},
 Journal = {{Journal of Applied Crystallography}},
-Year = {{1995}},
+Year = {1995},
 Volume = {{28}},
 Number = {{6}},
 Pages = {{768-773}},
@@ -35,7 +35,7 @@ Author = {Affentranger, R and Tavernelli, I and Di Iorio, EE},
 Title = {{A novel Hamiltonian replica exchange MD protocol to enhance protein
    conformational space sampling}},
 Journal = {{Journal of Chemical Theory and Computation}},
-Year = {{2006}},
+Year = {2006},
 Volume = {{2}},
 Number = {{2}},
 Pages = {{217-228}},
@@ -51,7 +51,7 @@ Unique-ID = {{ISI:000236056600001}},
 Author = {Rozycki, Bartosz and Kim, Young C. and Hummer, Gerhard},
 Title = {{SAXS Ensemble Refinement of ESCRT-III CHMP3 Conformational Transitions}},
 Journal = {{STRUCTURE}},
-Year = {{2011}},
+Year = {2011},
 Volume = {{19}},
 Number = {{1}},
 Pages = {{109-116}},
@@ -70,7 +70,7 @@ Author = {Krzeminski, Mickael and Marsh, Joseph A. and Neale, Chris and Choy,
    Wing-Yiu and Forman-Kay, Julie D.},
 Title = {{Characterization of disordered proteins with ENSEMBLE}},
 Journal = {{Bioinformatics}},
-Year = {{2013}},
+Year = {2013},
 Volume = {{29}},
 Number = {{3}},
 Pages = {{398-399}},
@@ -89,7 +89,7 @@ Unique-ID = {{ISI:000314892000018}},
 Author = {Kabsch, W},
 Title = {{Solution for best rotatition to relat 2 sets of vectors}},
 Journal = {{Acta Crystallograpica Section A}},
-Year = {{1976}},
+Year = {1976},
 Volume = {{32}},
 Number = {{SEP1}},
 Pages = {{922-923}},
