### 2018-08-16 9b0288e
correct reference formatting
