+++ paper/paper.bib
@@ -7,7 +7,6 @@
     Number = {{2}},
     Pages = {{382-389}},
     Month = {{FEB}},
-    Type = {{Article}},
     DOI = {{10.1016/j.cpc.2011.09.010}},
     ISSN = {{0010-4655}},
     Author-Email = {{joseph.curtis@nist.gov}},
@@ -16,8 +15,8 @@
 }
 
 @article{Svergun95,
-Author = {Svergun, D and Barberato, C and Koch, MHJ},
-Title = {{CRYSOL - A program to evaluate x-ray solution scattering of biological
+Author = {Svergun, D. and Barberato, C. and Koch, M. H. J.},
+Title = {{CRYSOL - A program to evaluate X-ray solution scattering of biological
    macromolecules from atomic coordinates}},
 Journal = {{Journal of Applied Crystallography}},
 Year = {1995},
@@ -31,7 +30,7 @@ Unique-ID = {{ISI:A1995TP80300013}},
 }
 
 @article{Affentranger06,
-Author = {Affentranger, R and Tavernelli, I and Di Iorio, EE},
+Author = {Affentranger, R. and Tavernelli, I. and Di Iorio, E. E.},
 Title = {{A novel Hamiltonian replica exchange MD protocol to enhance protein
    conformational space sampling}},
 Journal = {{Journal of Chemical Theory and Computation}},
@@ -50,7 +49,7 @@ Unique-ID = {{ISI:000236056600001}},
 @article{Rozycki11,
 Author = {Rozycki, Bartosz and Kim, Young C. and Hummer, Gerhard},
 Title = {{SAXS Ensemble Refinement of ESCRT-III CHMP3 Conformational Transitions}},
-Journal = {{STRUCTURE}},
+Journal = {{Structure}},
 Year = {2011},
 Volume = {{19}},
 Number = {{1}},
@@ -86,8 +85,8 @@ Unique-ID = {{ISI:000314892000018}},
 }
 
 @article{ Kabsch76,
-Author = {Kabsch, W},
-Title = {{Solution for best rotatition to relat 2 sets of vectors}},
+Author = {Kabsch, W.},
+Title = {{Solution for best rotatation to relate 2 sets of vectors}},
 Journal = {{Acta Crystallograpica Section A}},
 Year = {1976},
 Volume = {{32}},
