### 2018-08-16 f76b0e0
correct typos, replace some words
