+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'idpflex: Analysis of intrinsically disordered proteins by comparing MDsimulations to Small Angle Scattering experiments'
+title: 'idpflex: Analysis of Intrinsically Disordered Proteins by Comparing Simulations to Small Angle Scattering Experiments'
 tags:
     - sans
     - neutron scattering
@@ -31,29 +31,65 @@ bibliography: paper.bib
 
 # Summary
 
-It is estimated that about 30\% of the eucariotic proteome consists of intrinsically disordered proteins (IDP’s), yet their presence in public structural databases is severely underrepresented.
-IDP’s adopt heterogeneous inter-converting conformations with similar probabilities, preventing resolution of structures with X-Ray diffraction techniques.
-Small angle scattering (SAS) probes the average features of the conformational ensemble, which can
-prove unsatisfactory when very different ensembles share nearly identical average features.
-To address this shortcomings, atomistic molecular dynamics (MD) simulations combined with enhanced sampling methods such as the
-Hamiltonian replica exchange method are specially suitable [@Affentranger06], since they probe extensive regions of the IDP's free-energy phase space.
-Hence, this type of simulations can produce physically meaningful conformations and offer a full-featured description of the conformational landscape when properly validated against available experimental SAS data.
-The python package idpflex clusters the 3D conformationals resulting from an MD simulation into a
-hierarchical tree, with protein substates taking the role of tree nodes.
-Alternatively, the conformational ensemble of structures can be generated by other means than MD, such as torsional sampling of the protein backbone [@Curtis12].
-This flexibility is possible because idpflex is agnostic to the ensemble generation method.
-In contrast to other methods [@Rozycki11], idpflex does not initially discard any conformation by labelling it as incompatible with the available experimental data.
-This data represents averages over the whole conformational ensemble, and using an average as the criterion by which to discard any specific conformation can lead to erroneous discarding decisions.
-Clustering is performed according to structural similarity between conformations, defined by the root mean square deviation algorithm [@Kabsch76].
-Alternatively, idpflex can cluster conformations according to an Euclidean distance in the space spanned by a set of structural properties such as radius of gyration and end-to-end distance.
-Calculation of SAS intensities [@Svergun95] for each substate allows quantitative comparison to SAS data, yielding the probability of the IDP to adopt one of the conformations of any specific substate.
-Arranging tens of thousands of conformations into (typically) less than ten substates provides the researcher with a manageable number of macro-conformations from which to derive meaningful conclusions regarding the conformational freedom of the IDP.
-In addition to clustering, idpflex can compute structural features for each substate such as radius of gyration, end-to-end distance, asphericity, solvent exposed surface area, contact maps, and secondary structure content.
-All these properties require atomistic detail, thus idpflex is applicable to the study of IDP's but not to the study of quaternary protein arrangements, where clustering of coarse-grain simulations become much more appropriate [@Rozycki11].
-idpflex is extensible to include additional structural properties of interest.
-In summary, idpflex integrates MD simulations with SAS experiments to obtain the conformational ensemble of IDP, a pertinent problem in structural biology.
+It is estimated that about 30\% of the eucariotic proteome consists of
+intrinsically disordered proteins (IDP’s), yet their presence in public
+structural databases is severely underrepresented.
+IDP’s adopt heterogeneous inter-converting conformations with similar
+probabilities, preventing resolution of structures with X-Ray
+diffraction techniques.
+Small angle scattering (SAS) probes the average features of the
+conformational ensemble, which can prove unsatisfactory when very different
+ensembles share nearly identical average features.
+To address this shortcoming, atomistic molecular dynamics (MD) simulations
[…]
