### 2018-08-16 #895 @jmborr
@arfon I have proof-read the paper and I have two issues I don't know how to resolve.
1. Affiliations (there are two) show up together in the same line. Is this the intended behavior?
2. Hyperlinks in the PDF are not functional but I assume this is the intended behavior.
