+++ requirements.txt
@@ -1,4 +1,5 @@
 numpy
+cython
 scipy
 six
 h5py
