+++ idpflex/__init__.py
@@ -4,4 +4,4 @@
 
 __author__ = """Jose Borreguero"""
 __email__ = 'borreguero@gmail.com'
-__version__ = '0.1.5'
+__version__ = '__version__ = '0.1.5''
