### 2018-08-17 538f2f7
Bump version: 0.1.5 → 0.1.6
