### 2018-08-20 #895 @arfon
> Affiliations (there are two) show up together in the same line. Is this the intended behavior?

Yes, this is expected.

> Hyperlinks in the PDF are not functional but I assume this is the intended behavior.

Are they not functional when rendering on GitHub, or not functional when downloading the PDF? If the former, this is simply GitHub's security policy.
