### 2018-08-20 #895 @jmborr
@arfon I was viewing the PDF in Github. I downloaded it and the links work and the links are fine 😄
