### 2018-08-24 #895 @arfon
> @arfon Everything seems in order. Do I have to take any other step or simply wait until I'm assigned a reviewer?

Yes, sorry for the delay. We're a little backed up with submissions currently and are still waiting for an editor to pick up this submission. In the meantime, if you have any suggestions for potential reviewers from [this list](https://bit.ly/joss-reviewers) feel free to mention them here.
