### 2018-08-24 #895 @jmborr
I would think of reviewers with python as their primary language a background in soft-matter and/or computational biology, with a hint of physics expertise. Some  suggestion:
- tacaswell
- krother
- khinsen
- andim
- afrubin
- a-slide
- conradstack
- semacu
