### 2018-09-17 #895 @arfon
:wave: @pjotrp @george-githinji @brainstorm - would one of you be willing to edit this submission for JOSS?
