### 2018-09-21 #895 @brainstorm
@luminita, would you like to review this JOSS paper for me? :)
