### 2018-10-08 #895 @percolator
What would such a review comprise? Would we run the software, make a
code.review or should we review a paper?

On Mon, Oct 8, 2018 at 2:31 AM Roman Valls Guimera <notifications@github.com>
wrote:

> @percolator <https://github.com/percolator> (or one of your students),
> would you mind helping out reviewing this proteomics/MD software?
>
