### 2018-10-09 #895 @brainstorm
@percolator Actually both, but it usually doesn't take more than 30 minutes to wrap it all up, see:

https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

The code review is mostly due diligence with unit testing, nothing like a deep full code audit.

Here you have an example checklist on what you should expect from the review:

https://github.com/openjournals/joss-reviews/issues/967

Hope that helps, feel free to inquire more :)
