### 2018-10-09 #1007 @whedon
[REVIEW]: idpflex: Analysis of Intrinsically Disordered Proteins by Comparing Simulations to Small Angle Scattering Experiments
**Submitting author:** @jmborr (<a href="http://orcid.org/0000-0002-0866-8158">Jose M. Borreguero</a>)
**Repository:** <a href="https://github.com/jmborr/idpflex.git" target ="_blank">https://github.com/jmborr/idpflex.git</a>
**Version:** 0.1.7
