### 2018-10-11 #1007 @gjeuken
@brainstorm
Hi Roman,
I have some comments and questions.

First of all, everything seems to be in good order, according to the checklist. I would like to see some more formalism in the method description, as it is quite general, but I'm not sure it's in the scope of the Journal. (Am I wrong here?)

* Their version number is 0.1.6 and not 0.1.5 as it is stated above.
* For the **statement of need** they do a good job on the paper but it is absent in the documentation.
* I was unable to run their example notebooks, most likely due to a path mixup, although I didn't check enough to be sure. The error I get is:
`FileNotFoundError: [Errno 2] No such file or directory: 'crysol': 'crysol'`
But they do have unit tests and they work ok.

Lastly, I'm not sure how deep I should go investigating the functionality of the software. If you could make this more clear to me I would be very grateful.

But as I said before, overall it looks very good.
