### 2018-10-12 #1007 @brainstorm
@jmborr, can you fix the issues @gjeuken encountered?

@arfon, as I see it, @gjeuken has very good points regarding more scope/expectations on the reviewing task itself, a good one to discuss with the editorial board perhaps? Thanks @gjeuken for rising the issue!
