### 2018-10-12 #1007 @jmborr
@brainstorm Sorry, it was my intention to start working on these issues today but had not the time. I will address them tomorrow.
