### 2018-10-12 #1007 @jmborr
@brainstorm @gjeuken Regarding the insertion of an **statement of need** in the documentation, I though of inserting the Summary paper I submitted here in JOSS. Is this admissible?
