### 2018-10-12 #1007 @jmborr
@gjeuken [crysol](https://www.embl-hamburg.de/biosaxs/crysol.html) is a third-party software to generate  X-ray patterns, given an arrangement of atomic positions. It is assumed to be installed under any of the directories of `PATH`.
What I can do is to include the output of crysol in helper repository [idpflex_data](https://github.com/jmborr/idpflex_data). The example notebook would then check for an installation of crysol in the user's computer and if a `Not-Found` error is raised, then the notebook will fetch the crysol output from the helper repository. How does that sound?
