### 2018-10-13 #1007 @brainstorm
@jmborr What Gustavo meant is that you should include the statement of need in the documentation **as well**, it's good where it is already, but would be better to have it in **both** places.
