+++ docs/references.bib
@@ -1,0 +1,98 @@
+@article{Curtis12,
+    Author = {Curtis, Joseph E. and Raghunandan, Sindhu and Nanda, Hirsh and Krueger, Susan},
+    Title = {{SASSIE: A program to study intrinsically disordered biological molecules and macromolecular ensembles using experimental scattering restraints}},
+    Journal = {{Computer Physics Communications}},
+    Year = {2012},
+    Volume = {{183}},
+    Number = {{2}},
+    Pages = {{382-389}},
+    Month = {{FEB}},
+    DOI = {{10.1016/j.cpc.2011.09.010}},
+    ISSN = {{0010-4655}},
+    Author-Email = {{joseph.curtis@nist.gov}},
+    Journal-ISO = {{Comput. Phys. Commun.}},
+    Unique-ID = {{ISI:000298531400019}}
+}
+
+@article{Svergun95,
+Author = {Svergun, D. and Barberato, C. and Koch, M. H. J.},
+Title = {{CRYSOL - A program to evaluate X-ray solution scattering of biological
+   macromolecules from atomic coordinates}},
+Journal = {{Journal of Applied Crystallography}},
+Year = {1995},
+Volume = {{28}},
+Number = {{6}},
+Pages = {{768-773}},
+Month = {{DEC 1}},
+DOI = {{10.1107/S0021889895007047}},
+ISSN = {{0021-8898}},
+Unique-ID = {{ISI:A1995TP80300013}},
+}
+
+@article{Affentranger06,
+Author = {Affentranger, R. and Tavernelli, I. and Di Iorio, E. E.},
+Title = {{A novel Hamiltonian replica exchange MD protocol to enhance protein
+   conformational space sampling}},
+Journal = {{Journal of Chemical Theory and Computation}},
+Year = {2006},
+Volume = {{2}},
+Number = {{2}},
+Pages = {{217-228}},
+Month = {{MAR-APR}},
+DOI = {{10.1021/ct050250b}},
+ISSN = {{1549-9618}},
+EISSN = {{1549-9626}},
+ResearcherID-Numbers = {{Tavernelli, Ivano/H-5204-2012}},
+Unique-ID = {{ISI:000236056600001}},
+}
+
+@article{Rozycki11,
+Author = {Rozycki, Bartosz and Kim, Young C. and Hummer, Gerhard},
+Title = {{SAXS Ensemble Refinement of ESCRT-III CHMP3 Conformational Transitions}},
+Journal = {{Structure}},
+Year = {2011},
+Volume = {{19}},
+Number = {{1}},
+Pages = {{109-116}},
+Month = {{JAN 12}},
+DOI = {{10.1016/j.str.2010.10.006}},
+ISSN = {{0969-2126}},
+ResearcherID-Numbers = {{Rozycki, Bartosz/B-7005-2009
+   Hummer, Gerhard/A-2546-2013}},
+ORCID-Numbers = {{Rozycki, Bartosz/0000-0001-5938-7308
+   Hummer, Gerhard/0000-0001-7768-746X}},
+Unique-ID = {{ISI:000286348000014}},
+}
+
+@article{Krzeminski13,
+Author = {Krzeminski, Mickael and Marsh, Joseph A. and Neale, Chris and Choy,
+   Wing-Yiu and Forman-Kay, Julie D.},
+Title = {{Characterization of disordered proteins with ENSEMBLE}},
+Journal = {{Bioinformatics}},
+Year = {2013},
+Volume = {{29}},
+Number = {{3}},
+Pages = {{398-399}},
+Month = {{FEB 1}},
[…]
