### 2018-10-15 #1007 @gjeuken
@jmborr
In the case of CRYSOL I think just a link to the software would suffice.
Your suggestion is of course better, but that's up to you.
