### 2018-10-15 91a2c04
Refs #72 correct hyperlink to references
