+++ docs/references.bib
@@ -7,7 +7,7 @@
     Number = {{2}},
     Pages = {{382-389}},
     Month = {{FEB}},
-    DOI = {{10.1016/j.cpc.2011.09.010}},
+    DOI = {10.1016/j.cpc.2011.09.010},
     ISSN = {{0010-4655}},
     Author-Email = {{joseph.curtis@nist.gov}},
     Journal-ISO = {{Comput. Phys. Commun.}},
@@ -24,7 +24,7 @@ Volume = {{28}},
 Number = {{6}},
 Pages = {{768-773}},
 Month = {{DEC 1}},
-DOI = {{10.1107/S0021889895007047}},
+DOI = {10.1107/S0021889895007047},
 ISSN = {{0021-8898}},
 Unique-ID = {{ISI:A1995TP80300013}},
 }
@@ -39,7 +39,7 @@ Volume = {{2}},
 Number = {{2}},
 Pages = {{217-228}},
 Month = {{MAR-APR}},
-DOI = {{10.1021/ct050250b}},
+DOI = {10.1021/ct050250b},
 ISSN = {{1549-9618}},
 EISSN = {{1549-9626}},
 ResearcherID-Numbers = {{Tavernelli, Ivano/H-5204-2012}},
@@ -55,7 +55,7 @@ Volume = {{19}},
 Number = {{1}},
 Pages = {{109-116}},
 Month = {{JAN 12}},
-DOI = {{10.1016/j.str.2010.10.006}},
+DOI = {10.1016/j.str.2010.10.006},
 ISSN = {{0969-2126}},
 ResearcherID-Numbers = {{Rozycki, Bartosz/B-7005-2009
    Hummer, Gerhard/A-2546-2013}},
@@ -74,7 +74,7 @@ Volume = {{29}},
 Number = {{3}},
 Pages = {{398-399}},
 Month = {{FEB 1}},
-DOI = {{10.1093/bioinformatics/bts701}},
+DOI = {10.1093/bioinformatics/bts701},
 ISSN = {{1367-4803}},
 ResearcherID-Numbers = {{Choy, Wing-Yiu/O-7608-2017
    }},
@@ -92,7 +92,7 @@ Year = {1976},
 Volume = {{32}},
 Number = {{SEP1}},
 Pages = {{922-923}},
-DOI = {{10.1107/S0567739476001873}},
+DOI = {10.1107/S0567739476001873},
 ISSN = {{0108-7673}},
 Unique-ID = {{ISI:A1976CD98100035}},
 }
