### 2018-10-15 fb9e958
Refs #72 added sphinx theme for travis build
