### 2018-10-15 3077322
Refs #72 Do not treat warnings as errors
