+++ .travis.yml
@@ -41,7 +41,7 @@ script:
         flake8;
     fi
   - if [[ "${TRAVIS_PYTHON_VERSION}" == "3.6" ]]; then
-        sphinx-build -nWT -b html docs docs/_build/html;
+        sphinx-build -nT -b html docs docs/_build/html;
     fi
 
 notifications:
