### 2018-10-15 1841d77
Refs #50 try setting minimum version numbers for numpy and cython
