+++ requirements.txt
@@ -1,5 +1,5 @@
-numpy
-cython
+cython>=0.19
+numpy>=0.19
 scipy
 six
 h5py
