### 2018-10-16 bd04ae1
Refs #50 increase minimum version numbers for numpy and cython
