### 2018-10-16 e9db615
Refs #50 increase minimum version numbers for numpy and cython
