### 2018-10-16 f349753
Refs #50 bound version of mdtraj
