### 2018-10-16 4a3449b
Refs #50 bound version of mdtraj to  <=1.8
