### 2018-10-16 3aff39b
Refs #50 bound versions of cython and numpy
