### 2018-10-16 d09b854
Refs #50 set different requirements for readthedocs build
