+++ requirements_readthedocs.txt
@@ -5,7 +5,6 @@ six
 h5py
 lmfit
 pandas
-mdtraj
 MDAnalysis
 tqdm
 sphinx
