### 2018-10-16 ad4d7d7
Refs #50 set different requirements for readthedocs build
