### 2018-10-16 3eeb797
Refs #50 set different requirements for readthedocs build
