### 2018-10-16 #1007 @jmborr
This PR dealt with the statement of need:
https://github.com/jmborr/idpflex/pull/74

Unfortunately, the documentation build in [readthedocs](https://readthedocs.org/projects/idpflex/builds/) is failing because of a known problem for one of the dependency packages ([mdtraj](https://github.com/mdtraj/mdtraj/issues/1119)) when building the documentation in a virtual environment. I am going to try building the documentation in readthedocs [with conda](https://docs.readthedocs.io/en/latest/conda.html), instead of virtualenv.
