### 2018-10-16 fde1a2c
Refs #75 updated HISTORY file
