### 2018-10-16 4190192
Refs #75 minor changest to the conda support
