+++ condaenv.yml
@@ -1,0 +1,22 @@
+name: idpflex
+channels:
+  - conda-forge
+  - omnia
+
+dependencies:
+  - cython
+  - numpy
+  - scipy
+  - six
+  - h5py
+  - lmfit
+  - pandas
+  - tqdm
+  - sphinx
+  - sphinx_rtd_theme
+  - sphinxcontrib-bibtex
+  - matplotlib
+  - nglview
+  - pathos
+  - mdtraj
+  - MDAnalysis
