### 2018-10-18 582e0ec
Refs #75 split notebook requirements from rest
