+++ requirements_readthedocs.txt
@@ -1,16 +1,0 @@
-cython
-numpy
-scipy
-six
-h5py
-lmfit
-pandas
-tqdm
-sphinx
-sphinx_rtd_theme
-sphinxcontrib-bibtex
-matplotlib
-nglview
-pathos
-mdtraj
-MDAnalysis
