+++ requirements.txt
@@ -12,5 +12,3 @@ sphinx
 sphinx_rtd_theme
 sphinxcontrib-bibtex
 matplotlib
-nglview
-pathos
