+++ notebooks/conda_requirements.yml
@@ -1,0 +1,6 @@
+channels:
+  - conda-forge
+
+dependencies:
+  - nglview
+  - pathos
