+++ condaenv.yml
@@ -16,7 +16,5 @@ dependencies:
   - sphinx_rtd_theme
   - sphinxcontrib-bibtex
   - matplotlib
-  - nglview
-  - pathos
   - mdtraj
   - MDAnalysis
