### 2018-10-18 933b1e0
Refs #75 remove notebook dependencies from conda environment
