+++ condaenv.yml
@@ -1,4 +1,5 @@
 name: idpflex
+
 channels:
   - conda-forge
   - omnia
