### 2018-10-18 #1007 @jmborr
@gjeuken @brainstorm This [PR](https://github.com/jmborr/idpflex/pull/76) dealt with building correctly the [documentation in readthedocs](https://idpflex.readthedocs.io/en/latest/). Now the statement-of-need is public :smiley:
