### 2018-10-30 #1007 @gjeuken
@brainstorm
Everything seems in good order, but I did not check the functionality box as I do not yet know the scope of the functionality review.
