### 2018-10-30 #1007 @brainstorm
@gjeuken According to https://joss.readthedocs.io/en/latest/review_criteria.html#functionality, you should check the core functionality, in other words whether the software generates the expected outputs given some inputs... with your proteomics knowledge, can you verify that the outputs match what the software claims to achieve?
