### 2018-10-30 #1007 @jmborr
@gjeuken I suggest:
1. running the unit tests in your machine. After all, automatic testing with Travis only covers a few configurations.
2. running the two jupyter notebooks. These cover almost all of the software functionality. Currently, some of the input is generated by third party software ([CRYSOL](https://www.embl-hamburg.de/biosaxs/crysol.html), [DSSP](http://www.cmbi.ru.nl/xssp/)). I can modify the notebooks to make sure that the notebooks will fetch the required input from the data repository in case you don't have this third party software installed.

Does this sound good?
