### 2018-10-30 #1007 @gjeuken
@brainstorm
Sorry for being confusing, but my question is whether or not I should evaluate the "correctness" of the output in some way, or how much I should go down that path. That would require me to evaluate the methods themselves and for such, I would need a better description of them and it would resemble a more traditional article review.
An alternative is just to check if the input and output are in the correct format.

@jmborr
I've run the unit test and they all passed.
Regarding point 2, I think you already did some of those changes, but I could be wrong. I'm still having problems running the notebooks past the CRYSOL input, but I'm positive I could solve them with some effort if this is the direction the review should go.

Again, I'm sorry if I am confusing and dragging this too much, but this is a different review system than what I'm used to in general.
