### 2018-10-30 #1007 @jmborr
@gjeuken I will solve the issue regarding DSSP much in the same way as I did with CRYSOL. Notebook `hiAPP_clustering.ipynb` is a system test in itself because the task is to find the combination of clusters that provide the best fit to the experimental X-Ray profile. In the notebook, the experimental X-Ray profile is actually composed of three particular clusters, so the mail goal of the notebook is to show that the fitting procedure that scans the clustering tree is able to recover those three clusters. The rest of the notebook is a showcase of what properties one can calculate on the identified clusters for better characterization of the protein ensemble of structures.
