### 2018-10-30 #1007 @jmborr
This [PR](https://github.com/jmborr/idpflex/pull/79) dealt with checking for DSSP. If not found, the notebook fetches the output files from the data repository.
