### 2018-11-22 #1007 @brainstorm
@jmborr Please do you have a moment to review @gjeuken input and so we can submit and archive the DOI for your JOSS submission? :) Great work so far, thanks everyone 👍
