### 2018-11-22 aefc7d1
More detail to the summary
