+++ paper/paper.bib
@@ -14,6 +14,16 @@
     Unique-ID = {{ISI:000298531400019}}
 }
 
+@article{ Michaud11,
+Author = {N. Michaud-Agrawal, E. J. Denning, T. B. Woolf, and O. Beckstein},
+Title = {{MDAnalysis: A Toolkit for the Analysis of Molecular Dynamics Simulations}},
+Journal = {{Journal of Computational Chemistry}},
+Year = {2011},
+Volume = {{32}},
+Pages = {{2319–2327}},
+DOI = {{10.1002/jcc.21787}},
+}
+
 @article{Svergun95,
 Author = {Svergun, D. and Barberato, C. and Koch, M. H. J.},
 Title = {{CRYSOL - A program to evaluate X-ray solution scattering of biological
@@ -76,11 +86,6 @@ Pages = {{398-399}},
 Month = {{FEB 1}},
 DOI = {{10.1093/bioinformatics/bts701}},
 ISSN = {{1367-4803}},
-ResearcherID-Numbers = {{Choy, Wing-Yiu/O-7608-2017
-   }},
-ORCID-Numbers = {{Choy, Wing-Yiu/0000-0002-2124-7278
-   Forman-Kay, Julie/0000-0001-8265-972X
-   Marsh, Joseph/0000-0003-4132-0628}},
 Unique-ID = {{ISI:000314892000018}},
 }
 
@@ -88,11 +93,49 @@ Unique-ID = {{ISI:000314892000018}},
 Author = {Kabsch, W.},
 Title = {{Solution for best rotatation to relate 2 sets of vectors}},
 Journal = {{Acta Crystallograpica Section A}},
-Year = {1976},
+Year = {{1976}},
 Volume = {{32}},
 Number = {{SEP1}},
 Pages = {{922-923}},
 DOI = {{10.1107/S0567739476001873}},
 ISSN = {{0108-7673}},
 Unique-ID = {{ISI:A1976CD98100035}},
+}
+
+@article{ Gowers2016,
+Author = {R. J. Gowers et al.},
+Title = {{MDAnalysis: A Python package for the rapid analysis of molecular dynamics simulations}},
+Journal = {{Proceedings of the 15th Python in Science Conference}},
+Year = {{2016}},
+Pages = {{98-105}},
+}
+
+@article{ Kabsch83,
+author = {Kabsch, Wolfgang and Sander, Christian},
+title = {{Dictionary of protein secondary structure: Pattern recognition of hydrogen-bonded and geometrical features}},
+journal = {{Biopolymers}},
+Year = {{1983}},
+volume = {{22}},
+number = {{12}},
+pages = {{2577-2637}},
+doi = {{10.1002/bip.360221211}},
+DOI = {{10.1002/bip.360221211}},
+}
+
+@article{ McGibbon15,
+Author = {McGibbon, Robert T. and Beauchamp, Kyle A. and Harrigan, Matthew P. and
+   Klein, Christoph and Swails, Jason M. and Hernandez, Carlos X. and
+   Schwantes, Christian R. and Wang, Lee-Ping and Lane, Thomas J. and
+   Pande, Vijay S.},
+Title = {{MDTraj: A Modern Open Library for the Analysis of Molecular Dynamics
+   Trajectories}},
+Journal = {{BIOPHYSICAL JOURNAL}},
+Year = {{2015}},
+Volume = {{109}},
+Number = {{8}},
+Pages = {{1528-1532}},
+Month = {{OCT 20}},
+DOI = {{10.1016/j.bpj.2015.08.015}},
+ISSN = {{0006-3495}},
+EISSN = {{1542-0086}},
 }
