+++ docs/index.rst
@@ -19,56 +19,88 @@ intrinsically disordered proteins (IDP’s), yet their presence in public
 structural databases is severely underrepresented.
 IDP’s adopt heterogeneous inter-converting conformations with similar
 probabilities, preventing resolution of structures with X-Ray
-diffraction techniques.
-Small angle scattering (SAS) probes the average features of the
-conformational ensemble, which can prove unsatisfactory when very different
-ensembles share nearly identical average features.
+diffraction techniques. An alternative technique with wide application on
+IDP systems is small angle scattering (SAS). SAS can measure average
+structural features of IDP's when in vitro solution, or even at conditions
+mimicking protein concentrations found in the cell's cytoplasm.
+
+Despite these advantages, the averaging nature of SAS measurements
+will prove unsatisfactory if one aims to differentiate among the different
+conformations that a particular IDP can adopt. Different distributions
+of conformations can yield the same average therefore it is not possible to
+retrace the true distribution if all that SAS provides is the average
+conformation.
+
 To address this shortcoming, atomistic molecular dynamics (MD) simulations
-combined with enhanced sampling methods such as the Hamiltonian replica
-exchange method are specially suitable :cite:`Affentranger06`, since they
-probe extensive regions of the IDP's free-energy phase space.
-Hence, this type of simulations can produce physically meaningful
-conformations and offer a full-featured description of the
-conformational landscape when properly validated against available
-experimental SAS data.
-The python package idpflex clusters the 3D conformations resulting from
-an MD simulation into a hierarchical tree, with protein substates taking
-the role of tree nodes.
-Alternatively, the conformational ensemble of structures can be
-generated by other means than MD, such as torsional sampling of the
-protein backbone :cite:`Curtis12`.
-This flexibility is possible because idpflex is agnostic to the ensemble
-generation method.
-In contrast to other methods :cite:`Rozycki11`, idpflex does not initially
+of IDP systems combined with enhanced sampling methods such as the
+Hamiltonian replica exchange method are specially
+suitable :cite:`Affentranger06`. These simulations can probe extensive
+regions of the IDP's conformational space and have the potential to
+offer a full-featured description of the conformational landscape of IDP's.
+The results of these simulations should not be taken at faith value, however.
+First, a proper comparison against available experimental SAS data
+is a must. This validation step is the requirement that prompted the
+development of `idpflex`.
+
+The python package `idpflex` clusters the 3D conformations resulting from
+an MD simulation into a hierarchical tree by means of structural similarity
+among pairs of conformations. The conformations produced by the simulation
+take the role of Leafs in the hierarchichal tree. Nodes in the tree take the
+rol of IDP substates, with conformations under a particular Node making up
+one substate. Strictly speaking, `idfplex` does not require the
+IDP conformations to be produced by an MD simulation. Alternative conformation
+generators can be used, such as torsional sampling of the protein
+backbone :cite:`Curtis12`.
+In contrast to other methods :cite:`Rozycki11`, `idpflex` does not initially
 discard any conformation by labelling it as incompatible with the
-available experimental data.
-The data represent an average over all conformations, and using an
-average as the criterion by which to discard any specific conformation
-can lead to erroneous discarding decisions.
-Clustering is performed according to structural similarity between
-conformations, defined by the root mean square deviation
-algorithm :cite:`Kabsch76`.
-Alternatively, idpflex can cluster conformations according to an
-Euclidean distance in the space spanned by a set of structural
-properties such as radius of gyration and end-to-end distance.
-Calculation of SAS intensities :cite:`Svergun95` for each substate allows
-quantitative comparison to SAS data, yielding the probability of the
[…]

[…]
