### 2018-11-22 e1910a8
remove duplicate DOI field in one reference
