+++ paper/paper.bib
@@ -117,7 +117,6 @@ Year = {{1983}},
 volume = {{22}},
 number = {{12}},
 pages = {{2577-2637}},
-doi = {{10.1002/bip.360221211}},
 DOI = {{10.1002/bip.360221211}},
 }
 
