### 2018-11-22 #1007 @jmborr
@brainstorm @gjeuken I have split the summary paper into several paragraphs, adding more details. Please let me know how it looks now (https://idpflex.readthedocs.io/en/latest/)
