### 2018-11-22 #1007 @brainstorm
@jmborr Muchas gracias! ;)

@gjeuken To be clear, there's no need for you to fix the issues, but reporting them is a must... can you run the notebook successfully now?
