### 2018-12-04 #1007 @gjeuken
Hi all, sorry about the delay.
I still haven't manage to run the entire Notebook. the specific error I get is the following:
```
NameError                                 Traceback (most recent call last)
<ipython-input-12-8bed65896dd2> in <module>
----> 1 [profile.to_ascii(name) for profile, name in zip(profiles, crysol_names)]
      2 propagator_size_weighted_sum(profiles, cl.tree)

NameError: name 'profiles' is not defined
```

Which happens because the profile object is created in else clause that is never executed.

But just to be clear, my main question regarding this review was how much in detail should I look at the methods, and that is because I am used to reviewing papers for scientific magazines, where the mathematical method is the crucial in the article.
To answer my question I took the liberty of looking at other accepted articles here and it seams like this paper is not lacking in content in that respect. So I am sorry for causing this unecessary delay.
