### 2018-12-05 #1007 @Fahima-Islam
@gjeuken No need to be sorry, I'm very grateful you found a bug in the notebook! I have just pushed the fix to master.
Again, thanks for running the notebook :smiley:
