### 2018-12-06 #1007 @gjeuken
@jmborr
Perfect!
On the hiAPP_clustering_with_properties notebook you need to
`from distutils.spawn import find_executable`
And both of them work perfectly.

@brainstorm
With the notebooks now running, I was able to verify the complete functionality of the software, and I believe that was the only thing still remaining. :)

To both, again, sorry if I caused any delay in the process due to my confusion about the methods.
