### 2018-12-06 #1007 @jmborr
@gjeuken oops, that was sloppy from my part, sorry :blush:
I just pushed the fix to master.
