### 2018-12-07 #1007 @brainstorm
Thanks again for your dedication @gjeuken and @jmborr, nothing to be sorry about, good job!

@jmborr Can you let me know when you have a release ready with all the (known) bugs fixed?
