### 2018-12-07 3fda661
Bump version: 0.1.6 → 0.1.7
