+++ idpflex/__init__.py
@@ -4,4 +4,4 @@
 
 __author__ = 'Jose Borreguero'
 __email__ = 'borreguero@gmail.com'
-__version__ = '0.1.6'
+__version__ = '0.1.7'
