### 2018-12-07 #1007 @jmborr
I have uploaded the new version 0.1.7 with all the corrections to [pipy](https://pypi.org/project/idpflex/), and the [corrected documentation](https://idpflex.readthedocs.io/en/stable/)
