### 2018-12-10 #1007 @brainstorm
That's great, thanks @jmborr, can you deposit a release tarball on zenodo and share the DOI here?
