### 2018-12-10 #1007 @jmborr
@brainstorm never uploaded to *zenodo*, I am confused. Here's a screenshot for uploading the tarball. I assume the *Upload type* is **software** but I'm confused regarding the DOI. Should I click the *Reserve DOI* button?
![zenodo](https://user-images.githubusercontent.com/1136006/49738234-df661480-fc5c-11e8-940f-bef063010285.PNG)
