### 2018-12-10 #1007 @brainstorm
Yes, have a look at the other closed review issues in this repo, i.e: https://github.com/openjournals/joss-reviews/issues/1102#issuecomment-442090116, reserving a zenodo DOI is part of the process.
