### 2018-12-10 #1007 @brainstorm
@jmborr After reviewing the paper a bit more closely, it seems that there's a fair bit of repetition between the abstract and the actual paper.md, is that intentional?
