### 2018-12-10 fc806d3
remove duplicate paragraph
