### 2018-12-10 #1007 @jmborr
@brainstorm nope, I just pushed to master the removal of the duplicated paragraph
