+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{Curtis12,
-    Author = {Joseph E. Curtis et al.},
+    Author = {Joseph E. Curtis},
     Title = {{SASSIE: A program to study intrinsically disordered biological molecules and macromolecular ensembles using experimental scattering restraints}},
     Journal = {{Computer Physics Communications}},
     Year = {2012},
@@ -15,7 +15,7 @@
 }
 
 @article{Svergun95,
-Author = {D. Svergun et al.},
+Author = {D. Svergun},
 Title = {{CRYSOL - A program to evaluate X-ray solution scattering of biological
    macromolecules from atomic coordinates}},
 Journal = {{Journal of Applied Crystallography}},
@@ -30,7 +30,7 @@ Unique-ID = {{ISI:A1995TP80300013}},
 }
 
 @article{Affentranger06,
-Author = {R. Affentranger et al},
+Author = {R. Affentranger},
 Title = {{A novel Hamiltonian replica exchange MD protocol to enhance protein
    conformational space sampling}},
 Journal = {{Journal of Chemical Theory and Computation}},
@@ -47,7 +47,7 @@ Unique-ID = {{ISI:000236056600001}},
 }
 
 @article{Rozycki11,
-Author = {Bartosz Rozycki et al.},
+Author = {B. Rozycki},
 Title = {{SAXS Ensemble Refinement of ESCRT-III CHMP3 Conformational Transitions}},
 Journal = {{Structure}},
 Year = {2011},
@@ -65,7 +65,7 @@ Unique-ID = {{ISI:000286348000014}},
 }
 
 @article{Krzeminski13,
-Author = {Mickael Krzeminski, et al.},
+Author = {M. Krzeminski},
 Title = {{Characterization of disordered proteins with ENSEMBLE}},
 Journal = {{Bioinformatics}},
 Year = {2013},
@@ -92,28 +92,28 @@ Unique-ID = {{ISI:A1976CD98100035}},
 }
 
 @article{ Michaud11,
-Author = {N. Michaud-Agrawal et al.},
+Author = {N. Michaud-Agrawal},
 Title = {{MDAnalysis: A Toolkit for the Analysis of Molecular Dynamics Simulations}},
 Journal = {{Journal of Computational Chemistry}},
-Year = {{2011}},
+Year = {2011},
 Volume = {{32}},
 Pages = {{2319–2327}},
 DOI = {{10.1002/jcc.21787}},
 }
 
 @article{ Gowers2016,
-Author = {R. J. Gowers et al.},
+Author = {R. J. Gowers},
 Title = {{MDAnalysis: A Python package for the rapid analysis of molecular dynamics simulations}},
 Journal = {{Proceedings of the 15th Python in Science Conference}},
-Year = {{2016}},
+Year = {2016},
 Pages = {{98-105}},
 }
 
 @article{ Kabsch83,
-author = {Wolfgang Kabsch, and Christian Sander},
+author = {W. Kabsch},
 title = {{Dictionary of protein secondary structure: Pattern recognition of hydrogen-bonded and geometrical features}},
 journal = {{Biopolymers}},
-Year = {{1983}},
+Year = {1983},
 volume = {{22}},
 number = {{12}},
 pages = {{2577-2637}},
@@ -121,11 +121,11 @@ DOI = {{10.1002/bip.360221211}},
 }
 
 @article{ McGibbon15,
-Author = {Robert T. McGibbon et al.},
+Author = {R. T. McGibbon},
 Title = {{MDTraj: A Modern Open Library for the Analysis of Molecular Dynamics
    Trajectories}},
[…]
