+++ paper/paper.bib
@@ -98,7 +98,7 @@ Journal = {{Journal of Computational Chemistry}},
 Year = {2011},
 Volume = {{32}},
 Pages = {{2319–2327}},
-DOI = {{10.1002/jcc.21787}},
+DOI = {10.1002/jcc.21787},
 }
 
 @article{ Gowers2016,
@@ -117,7 +117,7 @@ Year = {1983},
 volume = {{22}},
 number = {{12}},
 pages = {{2577-2637}},
-DOI = {{10.1002/bip.360221211}},
+DOI = {10.1002/bip.360221211},
 }
 
 @article{ McGibbon15,
@@ -130,7 +130,7 @@ Volume = {{109}},
 Number = {{8}},
 Pages = {{1528-1532}},
 Month = {{OCT 20}},
-DOI = {{10.1016/j.bpj.2015.08.015}},
+DOI = {10.1016/j.bpj.2015.08.015},
 ISSN = {{0006-3495}},
 EISSN = {{1542-0086}},
 }
