### 2018-12-11 #1007 @jmborr
@brainstorm OK, now the proof looks very much correct. Are following steps:
- issue `@whedon accept`
- Check the final proof
- issue `@whedon accept deposit=true`
??
