### 2018-12-11 #1007 @jmborr
@arfon @brainstorm @gjeuken Thanks a million for giving me a chance to publish in JOSS and for the effort to carry out this review!  💃
